push translating, language state to Editor
